임시로 shader.h랑 box 제외
