+++ boxhead/boxhead/Main.hpp
@@ -1,18 +1,5 @@
 #pragma once
-#include <OpenGL.hpp>
-#include <Pipeline.hpp>
-#include <VertexStream.hpp>
-#include <Blobs.hpp>
-#include <Utils.hpp>
-#include <BlobUtils.hpp>
-
 #include "Framework.hpp"
-#include "misc.h"
-#include "shader.h"
-#include "input.h"
-#include "box.h"
-#include "zombie.h"
-#include "line.h"
 
 GLvoid Render(GLvoid);
 GLvoid UpdateView(const int w, const int h);
